Fix: Use float instead of Decimal for JSON serialization
Decimal serializes to strings in JSON, but frontend expects numbers.
Changed daily_budget, default_bid, keyword_bid to float in CampaignResponse.

+++ campaign_models.py
@@ -202,9 +202,9 @@ class CampaignResponse(BaseModel):
     root_group: Optional[str] = None
     keyword_count: int
     keywords: Optional[List[CampaignKeywordResponse]] = None
-    daily_budget: Decimal
-    default_bid: Decimal
-    keyword_bid: Optional[Decimal] = None
+    daily_budget: float  # Changed from Decimal to ensure JSON serializes as number
+    default_bid: float   # Changed from Decimal to ensure JSON serializes as number
+    keyword_bid: Optional[float] = None  # Changed from Decimal
     bidding_strategy: str
     start_date: str
     status: str
